% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_mt_data.R
\name{delete_mt_data}
\alias{delete_mt_data}
\title{Delete MT Image Filenames}
\usage{
delete_mt_data(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{delete_kirby21_data}},
\code{modalities = "MT"} so it cannot be specified}
}
\value{
Nothing is returned
}
\description{
Delete the files for the MT images
}
