% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_mt_data.R
\name{download_mt_data}
\alias{download_mt_data}
\title{Download MT Image Filenames}
\usage{
download_mt_data(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{download_kirby21_data}},
\code{modalities = "MT"} so it cannot be specified}
}
\value{
Indicator if the data is downloaded.
}
\description{
Download the files for the MT images
}
