% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mt_filenames.R
\name{get_mt_filenames}
\alias{get_mt_filenames}
\title{Get MT Image Filenames}
\usage{
get_mt_filenames(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{get_image_filenames}},
\code{modalities = "MT"} so it cannot be specified}
}
\value{
Vector fo filenames
}
\description{
Return the filenames for the MT images
}
\examples{
get_mt_filenames()
}

